@@ -30,13 +30,13 @@
     "not op_mini all"
   ],
   "dependencies": {
-    "@babel/core": "7.17.8",
+    "@babel/core": "7.17.9",
     "@babel/plugin-transform-runtime": "7.17.0",
     "@babel/preset-env": "7.16.11",
     "@babel/preset-react": "7.16.7",
     "@babel/preset-typescript": "7.16.7",
-    "@babel/runtime-corejs3": "7.17.8",
-    "@pmmmwh/react-refresh-webpack-plugin": "0.5.4",
+    "@babel/runtime-corejs3": "7.17.9",
+    "@pmmmwh/react-refresh-webpack-plugin": "0.5.5",
     "autoprefixer": "10.4.4",
     "babel-loader": "8.2.4",
     "babel-plugin-transform-imports": "2.0.0",
@@ -45,40 +45,40 @@
     "clean-webpack-plugin": "4.0.0",
     "clsx": "1.1.1",
     "copy-webpack-plugin": "10.2.4",
-    "core-js": "3.21.1",
+    "core-js": "3.22.0",
     "css-loader": "6.7.1",
     "css-minimizer-webpack-plugin": "3.4.1",
     "dotenv": "16.0.0",
-    "fork-ts-checker-webpack-plugin": "7.2.1",
+    "fork-ts-checker-webpack-plugin": "7.2.6",
     "html-loader": "3.1.0",
     "html-webpack-plugin": "5.5.0",
     "lodash": "4.17.21",
     "mini-css-extract-plugin": "2.6.0",
     "postcss": "8.4.12",
     "postcss-loader": "6.2.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.6",
-    "react-refresh": "0.11.0",
-    "react-router-dom": "6.2.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-redux": "8.0.0-rc.1",
+    "react-refresh": "0.12.0",
+    "react-router-dom": "6.3.0",
     "redux": "4.1.2",
     "redux-devtools-extension": "2.13.9",
     "redux-saga": "1.1.3",
     "reselect": "4.1.5",
-    "sass": "1.49.9",
+    "sass": "1.50.0",
     "sass-loader": "12.6.0",
     "style-loader": "3.3.1",
     "terser-webpack-plugin": "5.3.1",
-    "webpack": "5.70.0",
+    "webpack": "5.72.0",
     "webpack-bundle-analyzer": "4.5.0",
     "webpack-cli": "4.9.2"
   },
   "devDependencies": {
     "@types/jest": "27.4.1",
-    "@types/lodash": "4.14.180",
-    "@types/react": "17.0.43",
-    "@types/react-dom": "17.0.14",
-    "@types/react-redux": "7.1.23",
+    "@types/lodash": "4.14.181",
+    "@types/react": "18.0.5",
+    "@types/react-dom": "18.0.1",
+    "@types/react-redux": "7.1.24",
     "babel-jest": "27.5.1",
     "eslint-config-yialo": "0.6.20",
     "gh-pages": "3.2.3",
@@ -87,7 +87,7 @@
     "nodemon": "2.0.15",
     "stylelint-config-yialo": "0.2.27",
     "typescript": "4.6.3",
-    "webpack-dev-server": "4.7.4"
+    "webpack-dev-server": "4.8.1"
   },
   "engines": {
     "node": "16"
