@@ -1,8 +1,16 @@
-import ReactDOM from 'react-dom';
+import * as React from 'react';
+import ReactDOM from 'react-dom/client';
 
 import { App } from './c_app';
 
 export const init = (): void => {
   const $rootNode = document.getElementById('app-root');
-  ReactDOM.render(<App />, $rootNode);
+
+  if (!$rootNode) {
+    return;
+  }
+
+  const appRoot = ReactDOM.createRoot($rootNode);
+  const app = <React.StrictMode><App /></React.StrictMode>;
+  appRoot.render(app);
 };
