@@ -1,19 +1,19 @@
 import * as React from 'react';
 
-type TThemeName = 'light' | 'dark';
-type TThemeDispatch = React.Dispatch<React.SetStateAction<TThemeName>>;
+type ThemeName = 'light' | 'dark';
+type TThemeDispatch = React.Dispatch<React.SetStateAction<ThemeName>>;
 
-const ThemeStateContext = React.createContext<TThemeName | undefined>(undefined);
+const ThemeStateContext = React.createContext<ThemeName | undefined>(undefined);
 const ThemeUpdaterContext = React
   .createContext<TThemeDispatch | undefined>(undefined);
 
 const THEME_LS_KEY = 'ui-theme';
 
 const getInitialTheme = () => {
-  let theme: TThemeName | null;
+  let theme: ThemeName | null;
 
   try {
-    theme = (window.localStorage.getItem(THEME_LS_KEY) as TThemeName | null);
+    theme = (window.localStorage.getItem(THEME_LS_KEY) as ThemeName | null);
 
     if (!theme) {
       theme = 'light';
@@ -25,8 +25,12 @@ const getInitialTheme = () => {
   return theme;
 };
 
-export const ThemeProvider: React.FC = ({ children }) => {
-  const [theme, setTheme] = React.useState<TThemeName>(getInitialTheme);
+type ThemeProviderProps = {
+  children: React.ReactNode;
+};
+
+export const ThemeProvider: React.FC<ThemeProviderProps> = ({ children }) => {
+  const [theme, setTheme] = React.useState<ThemeName>(getInitialTheme);
 
   return (
     <ThemeStateContext.Provider value={theme}>
@@ -59,7 +63,7 @@ export const useThemeToggle = () => {
   }
 
   const toggleTheme = React.useCallback(() => {
-    setTheme((prevTheme: TThemeName) => (prevTheme === 'dark' ? 'light' : 'dark'));
+    setTheme((prevTheme: ThemeName) => (prevTheme === 'dark' ? 'light' : 'dark'));
   }, [setTheme]);
 
   return toggleTheme;
