@@ -2,6 +2,10 @@ import * as React from 'react';
 
 import style from './self.scss';
 
-export const MainLayout: React.FC = ({ children }) => {
+type MainLayoutProps = {
+  children: React.ReactNode;
+};
+
+export const MainLayout: React.FC<MainLayoutProps> = ({ children }) => {
   return <main className={style.root}>{children}</main>;
 };
